backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
deepgram-sdk==3.2.7
loguru==0.7.0